requirements.txt
stim==1.14.0
PyMatching==2.2.1
numpy
pandas
absl-py
scipy
